src/api/requirements_api.txt
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
build==0.10.0
CacheControl==0.12.14
cachetools==5.3.2
cffi==1.16.0
chardet==5.2.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==42.0.2
distlib==0.3.8
dulwich==0.21.7
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.13.1
h11==0.14.0
html5lib==1.1
idna==3.6
importlib-metadata==7.0.1
installer==0.6.0
jaraco.classes==3.3.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.13.1
lockfile==0.12.2
msgpack==1.0.7
numpy==1.26.4
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pkginfo==1.9.6
platformdirs==4.2.0
pluggy==1.4.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pyproject-api==1.6.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-multipart==0.0.9
rapidfuzz==3.6.1
referencing==0.33.0
requests-toolbelt==0.10.1
rpds-py==0.17.1
scikit-learn==1.4.0
scipy==1.12.0
shellingham==1.5.4
sniffio==1.3.0
starlette==0.36.3
threadpoolctl==3.3.0
tomli==2.0.1
tomlkit==0.12.3
tox==4.12.1
trove-classifiers==2024.1.31
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.1
virtualenv==20.25.0
